+++ test/test_schemas.py
@@ -2,7 +2,7 @@ from copy import deepcopy
 from test.utils import as_param
 
 import pytest
-from jsonschema import RefResolver
+from jsonschema import RefResolver, ValidationError
 
 import schemathesis
 from schemathesis.exceptions import InvalidSchema
@@ -63,7 +63,7 @@ def test_resolving_multiple_files():
         "schemes": ["http"],
         "produces": ["application/json"],
         "paths": {
-            "teapot": {
+            "/teapot": {
                 "post": {
                     "parameters": [
                         {
@@ -72,7 +72,8 @@ def test_resolving_multiple_files():
                             "name": "user",
                             "required": True,
                         }
-                    ]
+                    ],
+                    "responses": {"200": {"description": "OK"}},
                 }
             }
         },
@@ -94,8 +95,9 @@ def test_resolving_multiple_files():
     }
 
 
+@pytest.mark.parametrize("validate_schema, expected_exception", ((False, InvalidSchema), (True, ValidationError)))
 @pytest.mark.parametrize("error_type", ("KeyError", "AttributeError", "RefResolutionError"))
-def test_schema_parsing_error(simple_schema, error_type):
+def test_schema_parsing_error(simple_schema, error_type, validate_schema, expected_exception):
     raw_schema = deepcopy(simple_schema)
     if error_type == "KeyError":
         raw_schema.pop("paths")
@@ -103,6 +105,6 @@ def test_schema_parsing_error(simple_schema, error_type):
         raw_schema["paths"] = {None: ""}
     elif error_type == "RefResolutionError":
         raw_schema["paths"]["/users"]["get"]["parameters"] = [as_param({"$ref": "#/definitions/SimpleIntRef"})]
-    schema = schemathesis.from_dict(raw_schema)
-    with pytest.raises(InvalidSchema, match="Schema parsing failed. Please check your schema."):
+    with pytest.raises(expected_exception):
+        schema = schemathesis.from_dict(raw_schema, validate_schema=validate_schema)
         list(schema.get_all_endpoints())